Fixes 186: Refactor test suites for running individual test files
