+++ pkg/models/suite_test.go
@@ -19,6 +19,10 @@ type ModelsSuite struct {
 	tx *gorm.DB
 }
 
+type RepositoryRpmSuite struct {
+	*ModelsSuite
+}
+
 // Not using seeds.RandomOrgId to avoid cycle dependency
 var orgIDTest = strconv.Itoa(rand.Intn(99999999))
 var accountIdTest = strconv.Itoa(rand.Intn(99999999))
@@ -67,6 +71,6 @@ func (s *ModelsSuite) TearDownTest() {
 	s.tx.Rollback()
 }
 
-func TestRunSuiteModels(t *testing.T) {
+func TestModelsSuite(t *testing.T) {
 	suite.Run(t, new(ModelsSuite))
 }
